[ty] Simplify tuple pattern narrowing implementation
